# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: CPSC Backend API
  description: REST API for CPSC Cornerstone Project
  version: 1.0.0
  contact:
    name: CPSC Team
    email: support@cpsc.com

servers:
  - url: http://localhost:8080
    description: Local Development Server

tags:
  - name: Test
    description: Test endpoints for API verification
  - name: Authentication
    description: User authentication and registration endpoints
  - name: Institutions
    description: Financial institution management endpoints

paths:
  /api/hello:
    get:
      tags:
        - Test
      summary: Get a hello message
      description: Returns a simple greeting message to verify API is running
      operationId: hello
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Hello from CPSC Backend API!"
                  status:
                    type: string
                    example: "success"

  /api/secure/profile:
    get:
      tags:
        - Test
      summary: Get user profile (Protected)
      description: Returns user profile information. Requires authentication.
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response with user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Welcome to your profile!"
                  email:
                    type: string
                    example: "user@example.com"
                  authenticated:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account in AWS Cognito. User will receive a verification code via email.
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User successfully registered. Check email for verification code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
        '400':
          description: Invalid request or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/confirm:
    post:
      tags:
        - Authentication
      summary: Confirm user signup
      description: Confirms a user's signup with the verification code sent to their email
      operationId: confirmSignUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmSignUpRequest'
      responses:
        '200':
          description: User confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmSignUpResponse'
        '400':
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/resend-code:
    post:
      tags:
        - Authentication
      summary: Resend verification code
      description: Resends the verification code to the user's email
      operationId: resendConfirmationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendCodeRequest'
      responses:
        '200':
          description: Verification code resent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendCodeResponse'
        '400':
          description: Invalid request or user already confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate a user
      description: Logs in a user and returns JWT tokens. User must be confirmed before logging in.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Incorrect credentials or user not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/institutions:
    post:
      tags:
        - Institutions
      summary: Create a new financial institution
      description: Creates a new institution with a starting balance for the authenticated user
      operationId: createInstitution
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstitutionRequest'
      responses:
        '201':
          description: Institution created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      tags:
        - Institutions
      summary: Get all institutions for the authenticated user
      description: Returns a list of all financial institutions for the logged-in user
      operationId: getInstitutions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of institutions to return (default 50, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: lastEvaluatedKey
          in: query
          description: Pagination token from previous response to get next page
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved institutions
          content:
            application/json:
              schema:
                type: object
                properties:
                  institutions:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionResponse'
                  nextToken:
                    type: string
                    description: Token to retrieve the next page of results (null if no more results)
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint

  schemas:
    SignUpRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "Test@1234"
    
    SignUpResponse:
      type: object
      properties:
        message:
          type: string
          example: "User registered successfully. Check your email for verification code."
        userSub:
          type: string
          example: "12345678-1234-1234-1234-123456789012"
        confirmed:
          type: string
          example: "false"
    
    ConfirmSignUpRequest:
      type: object
      required:
        - email
        - confirmationCode
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        confirmationCode:
          type: string
          example: "123456"
          description: 6-digit verification code from email
    
    ConfirmSignUpResponse:
      type: object
      properties:
        message:
          type: string
          example: "User confirmed successfully. You can now login."
    
    ResendCodeRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    
    ResendCodeResponse:
      type: object
      properties:
        message:
          type: string
          example: "Verification code resent to your email"
    
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "Test@1234"
    
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token for API requests
        idToken:
          type: string
          description: JWT ID token containing user identity
        refreshToken:
          type: string
          description: Token used to refresh access tokens
        expiresIn:
          type: string
          description: Token expiration time in seconds
        tokenType:
          type: string
          example: "Bearer"
    
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Incorrect username or password"
    
    CreateInstitutionRequest:
      type: object
      required:
        - institutionName
        - startingBalance
      properties:
        institutionName:
          type: string
          minLength: 1
          maxLength: 100
          example: "Chase Bank"
          description: Name of the financial institution
        startingBalance:
          type: number
          format: double
          example: 324.50
          description: Starting balance in the account
    
    InstitutionResponse:
      type: object
      properties:
        institutionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: Unique identifier for the institution
        institutionName:
          type: string
          example: "Chase Bank"
          description: Name of the financial institution
        startingBalance:
          type: number
          format: double
          example: 324.50
          description: Starting balance in the account
        createdAt:
          type: string
          format: date-time
          example: "2025-12-24T12:00:00Z"
          description: Timestamp when the institution was created
        userId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          description: Cognito user ID (sub claim) of the user who owns this institution